switch to cli-docs-tool for yaml docs generation
switch to cli-docs-tool and validate yamldocs
