+++ .dockerignore
@@ -2,8 +2,11 @@
 /cli/winresources/versioninfo.json
 /cli/winresources/*.syso
 /man/man*/
-/docs/yaml/gen/
+/docs/yaml/
+/docs/vendor/
+/docs/go.sum
 profile.out
 
 # top-level go.mod is not meant to be checked in
 /go.mod
+/go.sum
