+++ docs/.gitignore
@@ -1,0 +1,3 @@
+/vendor
+/yaml
+/go.sum
