+++ docs/go.mod
@@ -1,0 +1,13 @@
+module github.com/docker/cli/docs
+
+// dummy go.mod to avoid dealing with dependencies specific
+// to docs generation and not really part of the project.
+
+go 1.16
+
+//require (
+//	github.com/docker/cli v0.0.0+incompatible
+//	github.com/docker/cli-docs-tool v0.3.0
+//)
+//
+//replace github.com/docker/cli v0.0.0+incompatible => ../
