+++ docs/tools.go
@@ -1,0 +1,7 @@
+// +build tools
+
+package main
+
+import (
+	_ "github.com/docker/cli-docs-tool"
+)
