--- docs/yaml/Dockerfile
@@ -1,4 +1,0 @@
-FROM scratch
-COPY docs /docs
-# CMD cannot be nil so we set it to empty string
-CMD  [""]
